+++ src/components/FlawFormOwner/FlawFormOwner.vue
@@ -60,7 +60,7 @@ const onQueryChange = async (query: string) => {
 };
 
 const handleSuggestionClick = (fn: (args?: any) => void, user: ZodJiraUserAssignableType) => {
-  owner.value = toValue(user.name);
+  owner.value = toValue(user.name ?? user.accountId ?? null);
   results.value = [];
   nextTick(fn);
 };
@@ -92,7 +92,7 @@ const handleSuggestionClick = (fn: (args?: any) => void, user: ZodJiraUserAssign
       <template v-if="results.length > 0" #suggestions="{ abort }">
         <div
           v-for="user in results"
-          :key="user.name"
+          :key="user.accountId ?? user.name ?? undefined"
           class="item"
           @click.prevent.stop="handleSuggestionClick(abort, user)"
         >