+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased]
 ### Added
 * Add highlighting on OSIDB Bot edited fields (`OSIDB-4790`)
+* Remove OSIDB Bot highlighting on flaw state change (`OSIDB-4824`)
 
 ### Fixed
 * Revert CVSS Aegis suggestion removes all value (`OSIDB-4780`)